Adding filereader.h
